# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Employee and Branch Management API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for the Employee and Branch Management application."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Branch": {
        "type": "object",
        "required": [
          "id",
          "name",
          "address",
          "phone"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the branch",
            "example": "branch_123abc"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "description": "Name of the branch",
            "example": "Downtown Branch"
          },
          "address": {
            "type": "string",
            "minLength": 5,
            "maxLength": 100,
            "description": "Physical address of the branch",
            "example": "123 Main Street, Winnipeg, MB"
          },
          "phone": {
            "type": "string",
            "pattern": "^[0-9]{10}$",
            "description": "Phone number of the branch (10 digits)",
            "example": "2045551234"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type or code",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "The name field is required"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "phone"
                },
                "issue": {
                  "type": "string",
                  "example": "must be a valid phone number"
                }
              }
            },
            "description": "Detailed validation errors (optional)"
          }
        }
      },
      "Employee": {
        "type": "object",
        "required": [
          "id",
          "name",
          "position",
          "department",
          "email",
          "phone",
          "branchId",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the employee",
            "example": "emp_123abc"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "description": "Full name of the employee",
            "example": "Jane Smith"
          },
          "position": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "description": "Job position/title of the employee",
            "example": "Software Developer"
          },
          "department": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "description": "Department where the employee works",
            "example": "Engineering"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Employee's email address",
            "example": "jane.smith@company.com"
          },
          "phone": {
            "type": "string",
            "minLength": 7,
            "maxLength": 20,
            "description": "Employee's phone number",
            "example": "+1-555-123-4567"
          },
          "branchId": {
            "type": "string",
            "description": "ID of the branch where the employee works",
            "example": "branch_456def"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the employee record was created",
            "example": "2024-01-15T10:30:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the employee record was last updated",
            "example": "2024-01-20T14:45:00Z"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/branches": {
      "get": {
        "summary": "Retrieve a list of all branches",
        "tags": [
          "Branches"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all branches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Branch"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new branch",
        "tags": [
          "Branches"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "address",
                  "phone"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 50,
                    "example": "Downtown Branch"
                  },
                  "address": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 100,
                    "example": "123 Main Street, Winnipeg, MB"
                  },
                  "phone": {
                    "type": "string",
                    "pattern": "^[0-9]{10}$",
                    "example": "2045551234"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Branch created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/branches/{id}": {
      "get": {
        "summary": "Retrieve a specific branch by ID",
        "tags": [
          "Branches"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the branch"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved branch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "404": {
            "description": "Branch not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update an existing branch's information",
        "tags": [
          "Branches"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the branch"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "address",
                  "phone"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 50,
                    "example": "Downtown Branch"
                  },
                  "address": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 100,
                    "example": "123 Main Street, Winnipeg, MB"
                  },
                  "phone": {
                    "type": "string",
                    "pattern": "^[0-9]{10}$",
                    "example": "2045551234"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Branch updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Branch not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a branch by ID",
        "tags": [
          "Branches"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the branch to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Branch deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Branch deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Branch not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/employees/branch/{branchId}": {
      "get": {
        "summary": "Retrieve all employees from a specific branch",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "branchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the branch"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved employees from the branch",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Branch not found or no employees in this branch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}